move group_func3d to func3d
